+++ docs/release_notes.md
@@ -1,5 +1,10 @@
 ## DigiNode Tools Release Notes
 
+DigiNode Tools v0.8.1 - 2023-08-15
+- Fix: Improve handling of a failed Kubo download - will now completely skip DigiAsset Node install/upgrade in this situation
+- Fix: Added workaround for Kubo release glitch - automatically downloads v0.22.0 when v0.21.1 is detected
+- Change: Improved one-page setup instructions for Raspberry Pi to make it clearer
+
 DigiNode Tools v0.8.0 - 2023-08-14
 - New: Now supports installing a pre-release version of DigiByte Core with the --dgpre flag. Downgrade back to the release version with --dgbnopre
 - New: Add --help flag to DigiNode Setup script which describes all the available optional launch flags